+++ client/web/src/enterprise/site-admin/sidebaritems.ts
@@ -152,12 +152,6 @@ export const codyGroup: SiteAdminSideBarGroup = {
             exact: true,
         },
         {
-            label: 'Embeddings (NEW!)',
-            to: '/site-admin/embeddings-new',
-            exact: true,
-        },
-
-        {
             label: 'Configuration',
             to: '/site-admin/embeddings/configuration',
         },
